requirements.txt
# For Oracle Database support:
# see https://oracle.github.io/odpi/doc/installation.html#linux

aliyun-log-python-sdk==0.9.12
arrow==0.17.0
clickhouse-driver==0.2.9
croniter==1.4.1
cx-Oracle==7.3.0
datasize==1.0.0
DBUtils==3.0.3
func-timeout==4.3.5
funcsigs==1.0.2
httpie==3.2.1
influxdb==5.3.1
Jinja2==2.11.3
kafka-python3==3.0.0
MarkupSafe==2.0.1 # Higher versions can cause Jinja2 to import errors.
nanoid==2.0.0
paho-mqtt==1.6.1
parse-args==2.0.0
psutil==5.9.1
psycopg2==2.9.10
pycryptodomex==3.15.0
pykafka==2.8.0
pymongo==3.12.3
pymssql==2.2.11;platform_machine=="x86_64"
PyMySQL[rsa,ed25519]==1.0.2
pyodps==0.11.3.1
python-dateutil==2.8.2
python-memcached==1.62
PyYAML==6.0.2
redis==4.6.0
requests==2.28.1
retry2==0.9.4
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.2
timeout-decorator==0.5.0
toml==0.10.2
tomli-w==1.0.0
tomli==2.0.1
ujson==5.9.0
xmltodict==0.13.0
